+++ scripts/backends/vault.sh
@@ -1,24 +1,24 @@
 #!/usr/bin/env sh
 
 if [ "${QUIET}" = "false" ]; then
-    log 'DEPRECATED: Vault driver is going to be remove in the next major version. Use vals driver instead.'
+    log 'DEPRECATED: Vault backend is going to be remove in the next major version. Use vals backend instead.'
 fi
 
 _VAULT="${HELM_SECRETS_VAULT_PATH:-vault}"
 
 # shellcheck disable=SC2034
-_DRIVER_REGEX='!vault [A-z0-9][A-z0-9/*\.\_\-]*\#[A-z0-9*\.\_\-][A-z0-9*\.\_\-]*'
+_BACKEND_REGEX='!vault [A-z0-9][A-z0-9/*\.\_\-]*\#[A-z0-9*\.\_\-][A-z0-9*\.\_\-]*'
 
-# shellcheck source=scripts/drivers/_custom.sh
-. "${SCRIPT_DIR}/drivers/_custom.sh"
+# shellcheck source=scripts/backends/_custom.sh
+. "${SCRIPT_DIR}/backends/_custom.sh"
 
 _vault() {
     # shellcheck disable=SC2086
-    set -- ${SECRET_DRIVER_ARGS} "$@"
+    set -- ${SECRET_BACKEND_ARGS} "$@"
     $_VAULT "$@"
 }
 
-_custom_driver_get_secret() {
+_custom_backend_get_secret() {
     _type=$1
     _SECRET_PATH="${2%#*}"
     _SECRET_FIELD="${2#*#}"
@@ -29,12 +29,12 @@ _custom_driver_get_secret() {
 
     if ! _vault kv get -format="${_type}" -field="${_SECRET_FIELD}" "${_SECRET_PATH}"; then
         echo 'Error while get secret from vault!' >&2
-        echo vault kv get -format="${_type}" -field="${_SECRET_FIELD}" "${_SECRET_PATH}" "${SECRET_DRIVER_ARGS}" >&2
+        echo vault kv get -format="${_type}" -field="${_SECRET_FIELD}" "${_SECRET_PATH}" "${SECRET_BACKEND_ARGS}" >&2
         exit 1
     fi
 }
 
-_custom_driver_is_yaml() {
+_custom_backend_is_yaml() {
     _type=$1
     _SECRET_PATH="${2%#*}"
     _SECRET_FIELD="${2#*#}"