+++ tests/assets/custom-backend.sh
@@ -1,11 +1,11 @@
 #!/usr/bin/env sh
 
 # shellcheck disable=SC2034
-_DRIVER_REGEX='!vault [A-z0-9][A-z0-9/*\.\_\-]*\#[A-z0-9*\.\_\-][A-z0-9*\.\_\-]*'
+_BACKEND_REGEX='!vault [A-z0-9][A-z0-9/*\.\_\-]*\#[A-z0-9*\.\_\-][A-z0-9*\.\_\-]*'
 
-. "${HELM_SECRETS_SCRIPT_DIR}/drivers/_custom.sh"
+. "${HELM_SECRETS_SCRIPT_DIR}/backends/_custom.sh"
 
-_custom_driver_get_secret() {
+_custom_backend_get_secret() {
     _type=$1
     _SECRET=$2